Didn't change any code, just moved my inventory pages into components and refactored my inventory pages to use my components
